+++ spectral_connectivity/__init__.py
@@ -6,7 +6,14 @@ directed connectivity measures from time series data using multitaper methods.
 
 # flake8: noqa
 from spectral_connectivity.connectivity import Connectivity
-from spectral_connectivity.transforms import Multitaper, prepare_time_series
+from spectral_connectivity.transforms import (
+    Multitaper,
+    MultitaperParameters,
+    estimate_frequency_resolution,
+    estimate_n_tapers,
+    prepare_time_series,
+    suggest_parameters,
+)
 from spectral_connectivity.utils import get_compute_backend
 from spectral_connectivity.wrapper import multitaper_connectivity
 
@@ -23,7 +30,11 @@ except ImportError:
 __all__ = [
     "Connectivity",
     "Multitaper",
+    "MultitaperParameters",
     "multitaper_connectivity",
     "prepare_time_series",
+    "estimate_frequency_resolution",
+    "estimate_n_tapers",
+    "suggest_parameters",
     "get_compute_backend",
 ]